完成后台管理页面

+++ java/fang/src/main/java/com/yinglongyhy/fang/rest/admin/HouseInfoTableRestController.java
@@ -4,9 +4,17 @@ package com.yinglongyhy.fang.rest.admin;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.yinglongyhy.fang.entity.HouseInfo;
+import com.yinglongyhy.fang.entity.Label2houseInfo;
+import com.yinglongyhy.fang.entity.Lease;
+import com.yinglongyhy.fang.entity.Picture2houseInfo;
+import com.yinglongyhy.fang.exception.RestApiException;
 import com.yinglongyhy.fang.service.IHouseInfoService;
+import com.yinglongyhy.fang.service.ILabel2houseInfoService;
+import com.yinglongyhy.fang.service.ILeaseService;
+import com.yinglongyhy.fang.service.IPicture2houseInfoService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,12 +28,22 @@ import org.springframework.web.bind.annotation.*;
  * @author yinglongyhy
  * @since 2020-12-22
  */
+@Slf4j
 @RestController
 @RequestMapping("/admin/houseInfo")
 public class HouseInfoTableRestController {
 
     @Autowired
     private IHouseInfoService houseInfoService;
+
+    @Autowired
+    private ILabel2houseInfoService label2houseInfoService;
+
+    @Autowired
+    private IPicture2houseInfoService picture2houseInfoService;
+
+    @Autowired
+    private ILeaseService leaseService;
 
     @GetMapping("page")
     @ApiOperation(value = "", notes = "")
@@ -52,6 +70,12 @@ public class HouseInfoTableRestController {
     @DeleteMapping("{id}")
     @ApiOperation(value = "", notes = "")
     public ResponseEntity delete(@PathVariable("id") @ApiParam(value = "id") Long id) {
+        if (leaseService.count(new QueryWrapper<Lease>().eq("house_info", id)) > 0) {
+            log.warn("删除房源失败，当前房源正在被使用 : {}", id);
+            throw new RestApiException("HouseInfoTableRestController", "delete error", "删除房源失败，当前房源正在被使用");
+        }
+        label2houseInfoService.remove(new QueryWrapper<Label2houseInfo>().eq("house_info", id));
+        picture2houseInfoService.remove(new QueryWrapper<Picture2houseInfo>().eq("house_info", id));
         houseInfoService.remove(new QueryWrapper<HouseInfo>().eq("id", id));
         return new ResponseEntity<>(HttpStatus.OK);
     }
