+++ java/fang/src/main/java/com/yinglongyhy/fang/rest/admin/PictureTableRestController.java
@@ -3,10 +3,15 @@ package com.yinglongyhy.fang.rest.admin;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.yinglongyhy.fang.entity.Lease;
 import com.yinglongyhy.fang.entity.Picture;
+import com.yinglongyhy.fang.entity.Picture2houseInfo;
+import com.yinglongyhy.fang.exception.RestApiException;
+import com.yinglongyhy.fang.service.IPicture2houseInfoService;
 import com.yinglongyhy.fang.service.IPictureService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,12 +25,16 @@ import org.springframework.web.bind.annotation.*;
  * @author yinglongyhy
  * @since 2020-12-22
  */
+@Slf4j
 @RestController
 @RequestMapping("/admin/picture")
 public class PictureTableRestController {
 
     @Autowired
     private IPictureService pictureService;
+
+    @Autowired
+    private IPicture2houseInfoService picture2houseInfoService;
 
     @GetMapping("page")
     @ApiOperation(value = "", notes = "")
@@ -52,6 +61,10 @@ public class PictureTableRestController {
     @DeleteMapping("{id}")
     @ApiOperation(value = "", notes = "")
     public ResponseEntity delete(@PathVariable("id") @ApiParam(value = "id") Long id) {
+        if (picture2houseInfoService.count(new QueryWrapper<Picture2houseInfo>().eq("picture", id)) > 0) {
+            log.warn("删除图片失败，当前图片正在被使用 : {}", id);
+            throw new RestApiException("PictureTableRestController", "delete error", "删除图片失败，当前图片正在被使用");
+        }
         pictureService.remove(new QueryWrapper<Picture>().eq("id", id));
         return new ResponseEntity<>(HttpStatus.OK);
     }
